Publish first alpha version
This commit updates version codes of the code documentation.

This version is a first approximation of the tool. It covers the entire
initial scope of functionalities proposed. And it seems to work
correctly.

Some more features are pending, as a more manageable interface via
scriptable, without requiring interaction. In addition, I also have in
mind several performance improvements.

However, this small tool has already begun to be useful to me. I am
happy with this little project.