fix: Fix bug with global opt. from a non-git path
Set global option crashed when try to use from a directory that does not
contains a git repository.

The current version of gitpython library (i.e. 2.1.11) has a strong
dependence between the 'config_writer()' function and the Repo objects.
A Repo type object has to represent a really existing git repository. If
the user wants to edit the global git configuration from a working
directory that does not contain a repository (case included in the scope
of gitcher), config_writer() and therefore gitcher option 'g' will fail.

So, the gitpython library will be removed from imports and the git
commands will be played with operating system calls instead.