fix: Isolate mode options check
Interactive and fast mode share some option codes, but others are
exclusive of one of them. In addition, 'listen' function had to check
hardcoded if an input was an escape command.

The redefined 'check_opt' function solve all this problems. Can receive
candidate options of both modes and candidate escape commands. A simple
set of bool flags allows a complex behavior that securizes the option
validation of the program in a very elegant way.