feat: Isolate program keys into a container module
Program keys are hardcoded into various gitcher functions. This new
class propose a container to store the escape and command option keys.

This class must to be instanced on the init of the gitcher execution.
So, in addiction of the above, this module can contain a set with the
values of git relative context keys (i.e.: profile names, user names,
user emails and user signing keys).