fix: Fix a bug with first use of Gitcher
When you used Gitcher for the first time or removed your cherfile for
any reason, Gitcher caught this situation and offers you the possibility
of create a new cherfile. However, after the creation of the dictionary
module, this was created very soon in the execution and tried to load
the cherfile content before the check.

This commit solves this bug. To do it simple, always creates a cherfile
if this is unreachable.