fix: Improve profiles presentation when list
Using Gitcher in interactive mode, there two possible presentation
configurations. The first is in a table and the second is in a list. The
second is used only if the first is not viable (i.e.: there is not
enough terminal window size).

When the list presentation was used, there was not showed the profile
names, so their manipulation was very complicated.

This issue was suggested by Santiago Fernández González via email.